move read_quotes into infra
