+++ crates/autopilot/src/database/quotes.rs
@@ -1,11 +1,8 @@
 use {
     super::Postgres,
     anyhow::{Context, Result},
-    database::{byte_array::ByteArray, orders::Quote},
-    model::order::OrderUid,
     shared::maintenance::Maintaining,
     sqlx::types::chrono::{DateTime, Utc},
-    std::collections::HashMap,
 };
 
 impl Postgres {
@@ -18,24 +15,6 @@ impl Postgres {
         let mut ex = self.pool.acquire().await?;
         database::quotes::remove_expired_quotes(&mut ex, max_expiry).await?;
         Ok(())
-    }
-
-    /// Get quotes for all orders in the auction.
-    ///
-    /// Doens't guarantee that all orders have quotes.
-    pub async fn read_quotes(
-        &self,
-        orders: impl Iterator<Item = &OrderUid>,
-    ) -> Result<HashMap<OrderUid, Quote>> {
-        let mut ex = self.pool.acquire().await?;
-        let mut quotes = HashMap::new();
-        for order in orders {
-            let order_uid = ByteArray(order.0);
-            if let Some(quote) = database::orders::read_quote(&mut ex, &order_uid).await? {
-                quotes.insert(*order, quote);
-            }
-        }
-        Ok(quotes)
     }
 }
