+++ crates/autopilot/src/infra/database/quotes/mod.rs
@@ -1,6 +1,6 @@
 use {super::Database, crate::domain, model::order::OrderUid, std::collections::HashMap};
 
-pub mod dto;
+pub mod postgres;
 
 impl Database {
     pub async fn read_quotes(
@@ -9,7 +9,7 @@ impl Database {
     ) -> Result<HashMap<OrderUid, domain::Quote>, Error> {
         let mut quotes = HashMap::new();
         for (id, quote) in self.db.read_quotes(orders).await? {
-            quotes.insert(id, dto::into_domain(quote)?);
+            quotes.insert(id, postgres::into_domain(quote)?);
         }
         Ok(quotes)
     }
@@ -20,5 +20,5 @@ pub enum Error {
     #[error("failed to read data from database")]
     DbError(#[from] anyhow::Error),
     #[error(transparent)]
-    Conversion(#[from] dto::InvalidConversion),
+    Conversion(#[from] postgres::InvalidConversion),
 }
